refactor: refact view (rename to index)
